real fix of the whole issue
