+++ src/components/panel/annotations/AnnotationsView.tsx
@@ -38,7 +38,7 @@ const AnnotationsView: FC = () => {
 
 
   function getContent() {
-    if (Object.keys(matchedAnnotationsMap).length === 0) return <EmptyAnnotations />
+    if (!matchedAnnotationsMap || Object.keys(matchedAnnotationsMap).length === 0) return <EmptyAnnotations />
     if (annotationsMode === 'aligned') return <AlignAnnotationsList />
     if (annotationsMode === 'list') return <AnnotationsList />
   }
